Update externs codegen for newtypes
