+++ src/Language/PureScript/CodeGen/JS.hs
@@ -23,14 +23,12 @@ module Language.PureScript.CodeGen.JS (
     identNeedsEscaping
 ) where
 
-import Data.Maybe (catMaybes, fromJust, fromMaybe)
+import Data.Maybe (catMaybes, fromJust)
 import Data.Function (on)
 import Data.List (nub, (\\), delete, sortBy)
 
 import Control.Monad (replicateM, forM)
 import Control.Applicative
-
-import qualified Data.Map as M
 
 import Language.PureScript.Names
 import Language.PureScript.Declarations
@@ -407,49 +405,3 @@ binderToJs m e varName done (NamedBinder ident binder) = do
   return (JSVariableIntroduction (identToJs ident) (Just (JSVar varName)) : js)
 binderToJs m e varName done (PositionedBinder _ binder) =
   binderToJs m e varName done binder
-
--- |
--- Finds the value stored for a data constructor in the current environment.
--- This is a partial function, but if an invalid type has reached this far then
--- something has gone wrong in typechecking.
---
-lookupConstructor :: Environment -> Qualified ProperName -> (DataDeclType, ProperName, Type)
-lookupConstructor e ctor = fromMaybe (error "Data constructor not found") $ ctor `M.lookup` dataConstructors e
-
--- |
--- Checks whether a data constructor is the only constructor for that type, used
--- to simplify the check when generating code for binders.
---
-isOnlyConstructor :: Environment -> Qualified ProperName -> Bool
-isOnlyConstructor e ctor = numConstructors (ctor, lookupConstructor e ctor) == 1
-  where
-  numConstructors :: (Qualified ProperName, (DataDeclType, ProperName, Type)) -> Int
-  numConstructors ty = length $ filter (((==) `on` typeConstructor) ty) $ M.toList $ dataConstructors e
-  typeConstructor :: (Qualified ProperName, (DataDeclType, ProperName, Type)) -> (ModuleName, ProperName)
-  typeConstructor (Qualified (Just moduleName) _, (_, tyCtor, _)) = (moduleName, tyCtor)
-  typeConstructor _ = error "Invalid argument to isOnlyConstructor"
-
--- |
--- Checks whether a data constructor is for a newtype.
---
-isNewtypeConstructor :: Environment -> Qualified ProperName -> Bool
-isNewtypeConstructor e ctor = case lookupConstructor e ctor of
-  (Newtype, _, _) -> True
-  (Data, _, _) -> False
-
--- |
--- Checks the number of arguments a data constructor accepts.
---
-getConstructorArity :: Environment -> Qualified ProperName -> Int
-getConstructorArity e = go . (\(_, _, ctors) -> ctors) . lookupConstructor e
-  where
-  go :: Type -> Int
-  go (TypeApp (TypeApp f _) t) | f == tyFunction = go t + 1
-  go (ForAll _ ty _) = go ty
-  go _ = 0
-
--- |
--- Checks whether a data constructor has no arguments, for example, `Nothing`.
---
-isNullaryConstructor :: Environment -> Qualified ProperName -> Bool
-isNullaryConstructor e = (== 0) . getConstructorArity e
